Refactored ProductModels from inventories to products
