+++ petclinic-frontend/src/features/products/ProductsList.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, JSX } from 'react';
 import { getAllProducts } from '@/features/products/api/getAllProducts.ts';
 import './ProductList.css';
-import { ProductModel } from '@/features/inventories/models/ProductModels/ProductModel.ts';
+import { ProductModel } from '@/features/products/models/ProductModels/ProductModel';
 import Product from './components/Product';
 
 export default function ProductList(): JSX.Element {
