+++ petclinic-frontend/src/features/products/api/getProduct.ts
@@ -1,5 +1,5 @@
 import axiosInstance from '@/shared/api/axiosInstance.ts';
-import { ProductModel } from '@/features/inventories/models/ProductModels/ProductModel.ts';
+import { ProductModel } from '@/features/products/models/ProductModels/ProductModel';
 
 export async function getProduct(productId: string): Promise<ProductModel> {
   const res = await axiosInstance.get('/products/' + productId, {
