+++ petclinic-frontend/src/features/products/models/ProductModels/RatingModel.ts
